+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
 * [Support] Drop support for PHP 5.6, to make use of new PHP 7.0 features.
 * [Support] Drop support for HHVM as there is little demand to make it worthwhile.
 * [Network] The construction of a new instance now only accepts a DeviceCollection instance.
+* [Network] Replaced the doctrine/cache dependency with any PSR-16 compatible cache library.
 * [Controllers] An exception is now throw when adding track(s) fails.
 * Scalar parameter type hints have been added where possible.
 * Return type hints have been added where possible.